moved model to shared
