+++ __test__/model/hand.legal.plays.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from '@jest/globals'
 import { createHand} from '../utils/test_adapter'
-import { Hand } from '../../src/model/hand'
+import { Hand } from '../../shared/model/hand'
 import { shuffleBuilder } from '../utils/shuffling'
 
 describe("Legal plays", () => {
