+++ __test__/model/hand.start.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, jest } from '@jest/globals'
 import { createHand} from '../utils/test_adapter'
-import { Card } from '../../src/model/deck'
-import { Hand } from '../../src/model/hand'
+import { Card } from '../../shared/model/deck'
+import { Hand } from '../../shared/model/hand'
 import { createHandWithShuffledCards, shuffleBuilder, successiveShufflers } from '../utils/shuffling'
 
 const normalShuffle = shuffleBuilder()
