+++ __test__/utils/shuffling.ts
@@ -1,6 +1,6 @@
-import { Card } from "../../src/model/deck"
-import { Hand } from "../../src/model/hand"
-import { Shuffler, standardShuffler } from "../../src/utils/random_utils"
+import { Card } from "../../shared/model/deck"
+import { Hand } from "../../shared/model/hand"
+import { Shuffler, standardShuffler } from "../../shared/utils/random_utils"
 import { CardPredicate, CardSpec, is, not } from "./predicates"
 import { HandProps, createHand, createInitialDeck } from "./test_adapter"
