+++ server/src/routes/gameRoutes.ts
@@ -4,11 +4,11 @@ import { gameController } from '../controllers/gameController';
 
 const router = express.Router();
 
-router.post('/games', gameController.createGame);
-router.post('/games/:id/join', gameController.joinGame);
-router.post('/games/:id/start', gameController.startGame);
-router.post('/games/:id/play', gameController.playTurn);
-router.get('/games/:id', gameController.getGameState);
-router.get('/games/', gameController.getGameList);
+router.post('/create', gameController.createGame);
+router.post('/:id/join', gameController.joinGame);
+router.post('/:id/start', gameController.startGame);
+router.post('/:id/play', gameController.playTurn);
+router.get('/:id', gameController.getGameState);
+router.get('/', gameController.getGameList);
 
 export default router;
